test: update tests to work with module structure
