+++ test/doc_tests.jl
@@ -1,0 +1,5 @@
+@testitem "Doctests" begin
+    # using Test
+    using Documenter
+    @testset doctest(HerbBenchmarks, manual=false)
+end
